+++ src/test/java/org/opensearch/neuralsearch/processor/NeuralQueryEnricherProcessorIT.java
@@ -21,7 +21,6 @@ import org.opensearch.common.xcontent.XContentHelper;
 import org.opensearch.common.xcontent.XContentType;
 import org.opensearch.core.rest.RestStatus;
 import org.opensearch.neuralsearch.BaseNeuralSearchIT;
-import org.opensearch.neuralsearch.query.HybridQueryBuilder;
 import org.opensearch.neuralsearch.query.NeuralQueryBuilder;
 import org.opensearch.neuralsearch.query.NeuralSparseQueryBuilder;
 
@@ -92,25 +91,25 @@ public class NeuralQueryEnricherProcessorIT extends BaseNeuralSearchIT {
         assertEquals(3, ((Map) responseInMap.get("hits")).size());
     }
 
-    @SneakyThrows
-    public void testNeuralQueryEnricherProcessor_whenHybridQueryBuilderAndNoModelIdPassed_thenSuccess() {
-        String modelId = null;
-        initializeIndexIfNotExist(index);
-        modelId = prepareModel();
-        createSearchRequestProcessor(modelId, search_pipeline);
-        createPipelineProcessor(modelId, ingest_pipeline, ProcessorType.TEXT_EMBEDDING);
-        updateIndexSettings(index, Settings.builder().put("index.search.default_pipeline", search_pipeline));
-        NeuralQueryBuilder neuralQueryBuilder = NeuralQueryBuilder.builder()
-            .fieldName(TEST_KNN_VECTOR_FIELD_NAME_1)
-            .queryText("Hello World")
-            .k(1)
-            .build();
-        HybridQueryBuilder hybridQueryBuilder = new HybridQueryBuilder();
-        hybridQueryBuilder.add(neuralQueryBuilder);
-        Map<String, Object> response = search(index, hybridQueryBuilder, 2);
-
-        assertFalse(response.isEmpty());
-    }
+    // @SneakyThrows
+    // public void testNeuralQueryEnricherProcessor_whenHybridQueryBuilderAndNoModelIdPassed_thenSuccess() {
+    // String modelId = null;
+    // initializeIndexIfNotExist(index);
+    // modelId = prepareModel();
+    // createSearchRequestProcessor(modelId, search_pipeline);
+    // createPipelineProcessor(modelId, ingest_pipeline, ProcessorType.TEXT_EMBEDDING);
+    // updateIndexSettings(index, Settings.builder().put("index.search.default_pipeline", search_pipeline));
+    // NeuralQueryBuilder neuralQueryBuilder = NeuralQueryBuilder.builder()
+    // .fieldName(TEST_KNN_VECTOR_FIELD_NAME_1)
+    // .queryText("Hello World")
+    // .k(1)
+    // .build();
+    // HybridQueryBuilder hybridQueryBuilder = new HybridQueryBuilder();
+    // hybridQueryBuilder.add(neuralQueryBuilder);
+    // Map<String, Object> response = search(index, hybridQueryBuilder, 2);
+    //
+    // assertFalse(response.isEmpty());
+    // }
 
     @SneakyThrows
     private void initializeIndexIfNotExist(String indexName) {